data_structure_game/requirements.txt
pygame==2.5.2
numpy==1.26.0
pyyaml==6.0.1
